+++ backend/src/main/resources/db/migration/internal/V0.123__update_env_actions_table.sql
@@ -4,16 +4,16 @@ ALTER TABLE public.missions
 
 CREATE TYPE mission_action_completion AS ENUM ('TO_COMPLETE', 'COMPLETED');
 
-ALTER TABLE public.env_actions ADD COLUMN action_completion mission_action_completion default 'TO_COMPLETE';
-ALTER TABLE public.env_actions ADD COLUMN action_open_by text;
-ALTER TABLE public.env_actions ADD COLUMN action_completed_by text;
+ALTER TABLE public.env_actions ADD COLUMN completion mission_action_completion default 'TO_COMPLETE';
+ALTER TABLE public.env_actions ADD COLUMN open_by text;
+ALTER TABLE public.env_actions ADD COLUMN completed_by text;
 
 
 -- Update `completion` column for past mission actions
 UPDATE env_actions
-SET action_completion = 'COMPLETED',
- action_open_by = missions.open_by,
- action_completed_by = missions.completed_by
+SET completion = 'COMPLETED',
+ open_by = missions.open_by,
+ completed_by = missions.completed_by
 FROM missions
 WHERE
     missions.id = env_actions.mission_id AND
