+++ datascience/tests/test_data/remote_database/V666.3__insert_dummy_env_actions.sql
@@ -4,7 +4,7 @@
 --
 
 INSERT INTO public.env_actions (
-                                         id, mission_id,    action_type, action_completion,                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         value, action_start_datetime_utc,   action_end_datetime_utc,                                                                                                   geom) VALUES
+                                         id, mission_id,    action_type, completion,                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         value, action_start_datetime_utc,   action_end_datetime_utc,                                                                                                   geom) VALUES
     ('dfb9710a-2217-4f98-94dc-283d3b7bbaae',         12, 'SURVEILLANCE', 'TO_COMPLETE',                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        '{}', '2022-11-20 20:31:41.719', '2022-11-20 23:31:41.719', ST_GeomFromText('MULTIPOLYGON(((-1.22 46.27,-1.22 46.27,-1.20 46.26,-1.19 46.26,-1.22 46.27)))', 4326)),
     ('d8e580fe-8e71-4303-a0c3-a76e1d4e4fc2',         12,      'CONTROL', 'COMPLETED',                                                                                                                                                                                                                                                                                                                                                                                            '{"infractions": [], "vehicleType": "VESSEL", "actionTargetType": "VEHICLE", "protectedSpecies": [], "actionNumberOfControls": 0, "actionStartDateTimeUtc": null}', '2022-11-24 20:31:41.719',                      NULL,                                   ST_GeomFromText('MULTIPOINT(-2.9822 48.1236,-3.0564 48.1177)', 4326)),
     ('88713755-3966-4ca4-ae18-10cab6249485',         19, 'SURVEILLANCE', 'TO_COMPLETE',                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                '{"observations": "Surveillance ok", "protectedSpecies": []}', '2022-11-28 13:59:20.176',  '2022-12-5 19:59:20.176',                               ST_GeomFromText('MULTIPOLYGON(((25 40,26 40,26 41,25 41,25 40)))', 4326)),
