+++ frontend/src/features/missions/MissionForm/utils.ts
@@ -1,4 +1,5 @@
-import { isEqual, omit } from 'lodash'
+import { saveMission } from 'domain/use_cases/missions/saveMission'
+import { debounce, isEmpty, isEqual, omit } from 'lodash'
 
 import { MISSION_EVENT_UNSYNCHRONIZED_PROPERTIES_IN_FORM } from './constants'
 import { isCypress } from '../../../utils/isCypress'
@@ -62,3 +63,33 @@ export function shouldSaveMission(
 function filterActionsFormInternalProperties(values: Partial<Mission | NewMission>) {
   return values.envActions?.map(envAction => omit(envAction, 'durationMatchesMission')) ?? []
 }
+
+export const validateBeforeOnChange = debounce(
+  async (
+    nextValues,
+    forceSave,
+    dispatch,
+    validateForm,
+    isAutoSaveEnabled,
+    engagedControlUnit,
+    selectedMission,
+    missionEvent
+  ) => {
+    const errors = await validateForm()
+    const isValid = isEmpty(errors)
+
+    if (!isAutoSaveEnabled || !isValid) {
+      return
+    }
+
+    if (!shouldSaveMission(selectedMission, missionEvent, nextValues) && !forceSave) {
+      return
+    }
+
+    if (engagedControlUnit) {
+      return
+    }
+    dispatch(saveMission(nextValues, false, false))
+  },
+  500
+)
